requirements.txt
flask==1.0.2
line-bot-sdk==1.8.0
